requirements.txt
sentence-transformers
faiss-cpu
openai
pdfplumber
ipywidgets
